% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_dataset.R
\docType{data}
\name{demo_md}
\alias{demo_md}
\title{Arabidopsis thaliana - Metadata dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 103 rows and 6 columns.
}
\source{
\doi{10.1038/nature11237}
}
\usage{
demo_md
}
\description{
This dataset was given by Lundberg et al., (2012). The metadata table contains
additional information about each sample (root) included in the study. It
typically includes details about the experimental conditions, environmental
factors, sample genotype, location, and other relevant contextual information.
Metadata is crucial for linking the microbial community data to specific
experimental variables and understanding how the root microbiome might vary in
response to different factors.
The original dataset contains 1049 samples(rows) and factors like soil type,
genotype, treatment, developmental stage and replication information in columns
for each particular sample.
}
\details{
Here only the portion of the dataset is taken for running the functions.
The dataset contains 103 rows and 6 columns.
}
\keyword{datasets}
