% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTU_dataset.R
\docType{data}
\name{demo_otu}
\alias{demo_otu}
\title{Arabidopsis thaliana - OTU dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 188 rows and 1440 columns.
}
\source{
\doi{10.1038/nature11237}
}
\usage{
demo_otu
}
\description{
This dataset was given by Lundberg et al., (2012). The OTU table is a central
part of the data set. It represents the abundance or presence/absence of
different microbial taxa (operational taxonomic units) in the root samples of
Arabidopsis thaliana. Each column in the OTU table corresponds to a specific
sample (root) from the study, and each row represents a different OTU, which
could be a species or a group of closely related organisms. The table contains
numerical values representing the count of each OTU in the corresponding samples.
The original dataset contains 18783 rows of OTUs and 1439 samples(columns).
}
\details{
Here only the portion of the dataset is taken for running the functions.

The dataset contains 188 rows and 1440 columns.
}
\keyword{datasets}
